# docs/swagger.yaml
basePath: /swagger/datahub-service
definitions:
  controllers.APIResponse:
    properties:
      data: {}
      msg:
        example: 操作成功
        type: string
      status:
        example: 0
        type: integer
    type: object
...
  controllers.DataSourceTestRequest:
    properties:
      config:
        additionalProperties: true
        description: 额外测试配置
        type: object
      data_source_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      test_type:
        description: connection, data_preview
        example: connection
        type: string
    required:
    - data_source_id
    - test_type
    type: object
  controllers.DataSourceTestResponse:
    properties:
      data:
        description: 预览数据
      duration:
        description: 测试耗时（毫秒）
        example: 250
        type: integer
      error:
        description: 错误信息
        example: ""
        type: string
      message:
        example: 连接成功
        type: string
      metadata:
        additionalProperties: true
        description: 元数据信息
        type: object
      success:
        example: true
        type: boolean
      suggestions:
        description: 优化建议
        example:
        - 检查网络连接
        - 验证数据库权限
        items:
          type: string
        type: array
      test_type:
        example: connection
        type: string
    type: object
...
  controllers.InterfaceTestRequest:
    properties:
      interface_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      options:
        additionalProperties: true
        description: 测试选项
        type: object
      parameters:
        additionalProperties: true
        description: 测试参数
        type: object
      test_type:
        description: data_fetch, performance, validation
        example: data_fetch
        type: string
    required:
    - interface_id
    - test_type
    type: object
  controllers.InterfaceTestResponse:
    properties:
      column_count:
        description: 字段数
        example: 8
        type: integer
      data:
        description: 返回数据
      data_types:
        additionalProperties:
          type: string
        description: 字段类型
        type: object
      duration:
        description: 调用耗时（毫秒）
        example: 150
        type: integer
      error:
        description: 错误信息
        example: ""
        type: string
      message:
        example: 接口调用成功
        type: string
      performance:
        additionalProperties: true
        description: 性能指标
        type: object
      row_count:
        description: 数据行数
        example: 25
        type: integer
      success:
        example: true
        type: boolean
      test_type:
        example: data_fetch
        type: string
      validation:
        additionalProperties: true
        description: 数据验证结果
        type: object
      warnings:
        description: 警告信息
        example:
        - 数据量较大，建议分页查询
        items:
          type: string
        type: array
    type: object
...
  controllers.UpdateBasicLibraryRequest:
    properties:
      description:
        example: 存储用户相关数据的基础库
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name_en:
        example: user_database
        type: string
      name_zh:
        example: 用户数据库
        type: string
      status:
        example: active
        type: string
    required:
    - id
    type: object
...
  models.BasicLibrary:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      data_sources:
        items:
          $ref: '#/definitions/models.DataSource'
        type: array
      description:
        example: 存储用户基础信息的数据库
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      interfaces:
        description: 关联关系
        items:
          $ref: '#/definitions/models.DataInterface'
        type: array
      name_en:
        example: user_basic_library
        type: string
      name_zh:
        example: 用户数据基础库
        type: string
      status:
        example: active
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.CleansingRule:
    properties:
      config:
        additionalProperties: true
        type: object
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      interface_id:
        type: string
      is_enabled:
        type: boolean
      order_num:
        type: integer
      type:
        type: string
    type: object
...
  models.DataInterface:
    properties:
      basic_library:
        allOf:
        - $ref: '#/definitions/models.BasicLibrary'
        description: 关联关系
      clean_rules:
        items:
          $ref: '#/definitions/models.CleansingRule'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      data_source:
        $ref: '#/definitions/models.DataSource'
      data_source_id:
        type: string
      description:
        type: string
      id:
        type: string
      interface_config:
        $ref: '#/definitions/models.JSONB'
      is_table_created:
        type: boolean
      library_id:
        type: string
      name_en:
        type: string
      name_zh:
        type: string
      parse_config:
        $ref: '#/definitions/models.JSONB'
      status:
        type: string
      table_fields_config:
        $ref: '#/definitions/models.JSONB'
      type:
        description: realtime, batch
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.DataSource:
    properties:
      basic_library:
        allOf:
        - $ref: '#/definitions/models.BasicLibrary'
        description: 关联关系
      category:
        description: stream, http, db, file
        type: string
      connection_config:
        $ref: '#/definitions/models.JSONB'
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      library_id:
        type: string
      name:
        type: string
      params_config:
        $ref: '#/definitions/models.JSONB'
      type:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.JSONB:
    additionalProperties: true
    type: object
...
  models.ScheduleConfig:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      data_interface:
        $ref: '#/definitions/models.DataInterface'
      data_source:
        allOf:
        - $ref: '#/definitions/models.DataSource'
        description: 关联关系
      data_source_id:
        description: 可选：数据源级别的调度
        type: string
      error_message:
        type: string
      id:
        type: string
      interface_id:
        description: 可选：接口级别的调度
        type: string
      is_enabled:
        type: boolean
      last_run_status:
        description: pending, running, success, failed
        type: string
      last_run_time:
        type: string
      next_run_time:
        type: string
      schedule_config:
        additionalProperties: true
        type: object
      schedule_type:
        description: cron, interval, manual
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
info:
  contact: {}
  description: 智慧园区数据底座后台服务，提供数据采集、处理、存储、治理和共享功能
  title: 数据底座服务 API
  version: "1.0"
paths:
...
  /basic-libraries/add-basic-library:
    post:
      consumes:
      - application/json
      description: 添加数据基础库
      parameters:
      - description: 数据基础库请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BasicLibrary'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 添加数据基础库
      tags:
      - 数据基础库
  /basic-libraries/add-datasource:
    post:
      consumes:
      - application/json
      description: 添加数据源
      parameters:
      - description: 数据源请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DataSource'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 添加数据源
      tags:
      - 数据基础库
  /basic-libraries/add-interface:
    post:
      consumes:
      - application/json
      description: 添加数据接口
      parameters:
      - description: 数据接口请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DataInterface'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 添加数据接口
      tags:
      - 数据基础库
  /basic-libraries/configure-schedule:
    post:
      consumes:
      - application/json
      description: 为批量数据源配置调度任务
      parameters:
      - description: 调度配置请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ScheduleConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 配置数据源调度
      tags:
      - 数据基础库
  /basic-libraries/datasource-status/{id}:
    get:
      description: 获取数据源的连接状态、最近同步时间等信息
      parameters:
      - description: 数据源ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取数据源运行状态
      tags:
      - 数据基础库
  /basic-libraries/delete-basic-library:
    post:
      consumes:
      - application/json
      description: 删除数据基础库
      parameters:
      - description: 数据基础库请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BasicLibrary'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 删除数据基础库
      tags:
      - 数据基础库
  /basic-libraries/delete-datasource:
    post:
      consumes:
      - application/json
      description: 删除数据源
      parameters:
      - description: 数据源请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DataSource'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 删除数据源
      tags:
      - 数据基础库
  /basic-libraries/delete-interface:
    post:
      consumes:
      - application/json
      description: 删除数据接口
      parameters:
      - description: 数据接口请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DataInterface'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 删除数据接口
      tags:
      - 数据基础库
  /basic-libraries/interface-preview/{id}:
    get:
      description: 获取接口的样例数据用于预览
      parameters:
      - description: 接口ID
        in: path
        name: id
        required: true
        type: string
      - default: 10
        description: 数据条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 预览接口数据
      tags:
      - 数据基础库
  /basic-libraries/test-datasource:
    post:
      consumes:
      - application/json
      description: 测试数据源连接和数据获取能力
      parameters:
      - description: 测试请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.DataSourceTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controllers.DataSourceTestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 测试数据源连接
      tags:
      - 数据基础库
  /basic-libraries/test-interface:
    post:
      consumes:
      - application/json
      description: 测试数据接口的调用和数据获取能力
      parameters:
      - description: 测试请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.InterfaceTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controllers.InterfaceTestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 测试接口调用
      tags:
      - 数据基础库
  /basic-libraries/update-basic-library:
    post:
      consumes:
      - application/json
      description: 修改数据基础库信息
      parameters:
      - description: 修改数据基础库请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateBasicLibraryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 修改数据基础库
      tags:
      - 数据基础库
...
swagger: "2.0"